+++ src/libs/blueprint/conduit_blueprint.hpp
@@ -23,8 +23,6 @@
 #include "conduit_blueprint_mesh_examples_julia.hpp"
 #include "conduit_blueprint_mesh_examples_venn.hpp"
 
-#include "conduit_blueprint_util_mesh.hpp"
-
 #include "conduit_blueprint_o2mrelation.hpp"
 #include "conduit_blueprint_o2mrelation_examples.hpp"
 #include "conduit_blueprint_o2mrelation_iterator.hpp"